[Command] Move ProjectDirectory into dedicated file
