+++ lib/cocoapods/command/project.rb
@@ -1,40 +1,8 @@
 module Pod
   class Command
-    # Provides support for commands to take a user-specified `project directory`
-    #
-    module ProjectDirectory
-      module Options
-        def options
-          [
-            ['--project-directory=/project/dir/', 'The path to the root of the project directory'],
-          ].concat(super)
-        end
-      end
-
-      def self.included(base)
-        base.extend(Options)
-      end
-
-      def initialize(argv)
-        if project_directory = argv.option('project-directory')
-          @project_directory = Pathname.new(project_directory).expand_path
-        end
-        config.installation_root = @project_directory
-        super
-      end
-
-      def validate!
-        super
-        if @project_directory && !@project_directory.directory?
-          raise Informative,
-                "`#{@project_directory}` is not a valid directory."
-        end
-      end
-    end
-
-    #-------------------------------------------------------------------------#
-
     class Install < Command
+      include ProjectDirectory
+
       self.summary = 'Install project dependencies to Podfile.lock versions'
 
       self.description = <<-DESC
@@ -76,6 +44,8 @@ module Pod
     #-------------------------------------------------------------------------#
 
     class Update < Command
+      include ProjectDirectory
+
       self.summary = 'Update outdated project dependencies and create new ' \
         'Podfile.lock'
